+++ src/entities/comment/api/endpoint.ts
@@ -1,0 +1,5 @@
+export const COMMENT_URL = {
+  base: '/comments',
+  comments: (id: string) => `${COMMENT_URL.base}/courses/${id}`,
+  detail: (id: string) => `${COMMENT_URL.base}/${id}`,
+}
