+++ src/entities/comment/api/index.ts
@@ -1,0 +1,6 @@
+export * from './endpoint'
+export * from './queryKey'
+export * from './get-comments'
+export * from './post-comment'
+export * from './delete-comment'
+export * from './update-comment'
