+++ src/entities/comment/api/queryKey.ts
@@ -1,0 +1,4 @@
+export const COMMENT_QUERY_KEY = {
+  all: ['comments'] as const,
+  detail: (id: string) => [...COMMENT_QUERY_KEY.all, id] as const,
+}
